+++ NEWS.md
@@ -1,11 +1,17 @@
-# AMR 0.8.0.9003
-<small>Last updated: 23-Oct-2019</small>
+# AMR 0.8.0.9004
+<small>Last updated: 26-Oct-2019</small>
+
+### New
+* Support for a new MDRO guideline: Magiorakos AP, Srinivasan A *et al.* "Multidrug-resistant, extensively drug-resistant and pandrug-resistant bacteria: an international expert proposal for interim standard definitions for acquired resistance." Clinical Microbiology and Infection (2012). **This is now the new default guideline for the `mdro()` function.**
 
 ### Changes
 * When running `as.rsi()` over a data set, it will now print the guideline that will be used if it is not specified by the user
 * Fix for `eucast_rules()`: *Stenotrophomonas maltophilia* not interpreted "R" to ceftazidime anymore (following EUCAST v3.1)
 * Fix in taxonomic info for genera that are in multiple kingdoms, like *Proteus*
 * Fix for interpreting MIC values with `as.rsi()` where the input is `NA`
+* Added "imi" as allowed abbreviation for Imipenem
+* Fix for automatically determining columns with antibiotic results in `mdro()` and `eucast_rules()`
+* Added ATC codes for ceftaroline, ceftobiprole and faropenem
 
 # AMR 0.8.0